+++ app/Http/Controllers/CustomerTypeController.php
@@ -18,7 +18,7 @@ class CustomerTypeController extends Controller
      */
     public function index(Request $request)
     {
-        abort_unless(Gate::allows('loan_access'), 404);
+        abort_unless(Gate::allows('loan_access') || Gate::allows('branch_access'), 404);
         $lists = CustomerType::with('user')
         ->where('code', 'LIKE', '%', $request->search, '%')->orderBy("created_at", "asc")
         ->get();
@@ -28,7 +28,7 @@ class CustomerTypeController extends Controller
 
     public function add()
     {
-        abort_unless(Gate::allows('loan_access'), 404);
+        abort_unless(Gate::allows('loan_access') || Gate::allows('branch_access'), 404);
         $collectors = User::where('roles.title', 'Collector')
         ->join('role_user', 'users.id', '=', 'role_user.user_id')
         ->join('roles', 'role_user.role_id', '=', 'roles.id')
@@ -55,7 +55,7 @@ class CustomerTypeController extends Controller
      */
     public function store(CustomerTypeCreateRequest $request)
     {
-        abort_unless(Gate::allows('loan_access'), 404);
+        abort_unless(Gate::allows('loan_access') || Gate::allows('branch_access'), 404);
         if($request->validated()){
             $customer = new CustomerType();
             $customer->code = $request->code;
@@ -75,7 +75,7 @@ class CustomerTypeController extends Controller
      */
     public function show($id)
     {
-        abort_unless(Gate::allows('loan_access'), 404);
+        abort_unless(Gate::allows('loan_access') || Gate::allows('branch_access'), 404);
         $customer = CustomerType::where('id', $id)->first();
         $collectors = User::where('roles.title', 'Collector')
         ->join('role_user', 'users.id', '=', 'role_user.user_id')
@@ -105,7 +105,7 @@ class CustomerTypeController extends Controller
      */
     public function update(CustomerTypeUpdateRequest $request, string $id)
     {
-        abort_unless(Gate::allows('loan_access'), 404);
+        abort_unless(Gate::allows('loan_access') || Gate::allows('branch_access'), 404);
         if($request->validated()){
             $customer = CustomerType::find($id);
             $customer->code = $request->code;
@@ -125,7 +125,7 @@ class CustomerTypeController extends Controller
      */
     public function destroy($id)
     {
-        abort_unless(Gate::allows('loan_access'), 404);
+        abort_unless(Gate::allows('loan_access') || Gate::allows('branch_access'), 404);
         $customer = CustomerType::find($id);
         $customer->delete();
